Video 23, 24 en 25.

+++ app/Http/Controllers/JobController.php
@@ -2,8 +2,10 @@
 
 namespace App\Http\Controllers;
 
+use App\Mail\JobPosted;
 use App\Models\Job;
-use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
+use Illuminate\Support\Facades\Mail;
 
 class JobController extends Controller
 {
@@ -33,11 +35,15 @@ class JobController extends Controller
             'salary' => ['required']
         ]);
 
-        Job::create([
+        $job = Job::create([
             'title' => request('title'),
             'salary' => request('salary'),
             'employer_id' => 1
         ]);
+
+        Mail::to($job->employer->user)->queue(
+            new JobPosted($job)
+        );
 
         return redirect('/jobs');
     }
@@ -45,12 +51,11 @@ class JobController extends Controller
     public function edit(Job $job)
     {
         return view('jobs.edit', ['job' => $job]);
-
     }
 
     public function update(Job $job)
     {
-        // authorize (On hold...)
+        Gate::authorize('edit-job', $job);
 
         request()->validate([
             'title' => ['required', 'min:3'],
@@ -67,7 +72,7 @@ class JobController extends Controller
 
     public function destroy(Job $job)
     {
-        // authorize (On hold...)
+        Gate::authorize('edit-job', $job);
 
         $job->delete();
